% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyfishies.R
\name{shinyfishies}
\alias{shinyfishies}
\title{shinyfishies}
\usage{
shinyfishies()
}
\value{
An unispiring histogram...
}
\description{
This Shiny App displays a histogram of the DDT data set
}
\examples{
\dontrun{shinyfishies()}
}
